1420533: Add no_proxy option to API, config, UI
Adding `no_proxy` to both config file and API makes it more consistent
with other proxy options, in that it can be specified in environment,
config, or by code, and lookup is consistent for all proxy options
(API preferred over config, config preferred over environment).

For subscription-manager. this results in the precedence as
command-line > config > environment, as with other proxy options.