1428435: Make release set/unset regenerate repos
Previously we were relying on subsequent yum commands to regenerate
redhat.repo, but this is problematic because if the lock can't be
acquired, due to behavior introduced in f9715503 (bz 1224806), the repo
is simply left alone.

This change makes it so that the release command is responsible for
updating the repo. Furthermore, the release command will block if the
lock is held, until it can be acquired (with default timeout, this
should be no longer than 3 minutes).