1435771: Allow rhsm dbus services to log to the default rhsm.log
This commit also stops the facts service from failing in the event
there is malformed or empty custom facts file(s).
There is historical precedent for a different process, one other
than subscription-manager, writing to the default log file.
While I do not think this is the best thing to do, it is certainly
better than not logging at all.

Future work: Improve multiprocess logging for the rhsm ecosystem.