1438561: Do not use D-Bus for facts collection
Instead, we use the same underlying service, which is also exposed via
D-Bus.

This involved refactoring the aggregation into a separate service -
`AllFactsCollector`, and then using that to implement `AllFacts`.

I also moved the default configuration of the collectors into their
classes.

Also, I removed now outdated comments and logging that suggested we are
using D-Bus for facts collection.