1435013: Add splay option to rhsmcertd, randomize over interval
This is a second implementation of the splay feature added to
rhsmcertd. This adds a 'splay' option. The new option is a bool
and can be set in the rhsm.conf and disabled via the new
'--no-splay' cli option. With this feature enabled, a random
offset in seconds is chosen for both cert checks and auto attaches.
There are different offsets chosen for each type of check.
The offsets chosen will always be from 0 to the interval amount
of time. For example, if the autoAttachInterval is set to 60
seconds, the offset for splay will be between 0 and 60 seconds.
Both initial checks will be delayed by the chosen amount. After
the initial delay, the checks will be performed on the intervals
set in the config (or on the cli). In this way, the peak load on
the server that a series of systems are all checking in to can be
diffused or spread out over the intervals.