1438869: Capture dmidecode errors at fact gathering
Previously tried to fix with #1584, but with that fix, only some of the
cases were covered. This change covers more cases by making `get_all`
responsible for importing dmidecode and then making sure any warnings
are cleared.

In this way, dmiinfo more fully encapsulates dmidecode, not even
exposing the module outside of that method (except to pass to other
helper functions).