1422196: Update container certs after plugin install
In order to accomplish this, I added an invocation of `python
content_plugins/container_content.py` to the post-install scriptlet of
`subscription-manager-plugin-container`.

I added the main method to `container_content.py`, as this kept the
logic of invoking the hook close to the hook itself, and avoided
complexities with adjusting the `PYTHONPATH`.

By invoking the hook (rather than say, calling rhsmcertd-worker), we
avoid going to the network, and instead simply reference entitlements
already present to update the client certs for registry access
(currently at /etc/docker/certs.d).

Testing requires an entitlement with a product with matching tags and
content of type "containerimage".

I changed `HOSTNAME_CERT_DIR` to be a class attribute so it could be
more easily mocked.