Adds a new vagrant box "opensuse42.2"
This commit adds a new vagrant box.
The new vagrant box may be created and used by the following command:
`vagrant up opensuse42.2`.
In addition to the new box this commit includes a few changes to some of the
vagrant configuration. All boxes now are given 1024mb of memory.
All build and build_ext folders for both subman and python-rhsm are no longer
rsynced to the boxes.