Add "watch" and "vagrant-watch" tasks
`watch` does `webpack --watch` which is suitable for running in any
environment.

`vagrant-watch` does both `webpack --watch` and `vagrant rsync-auto`
which is suitable for coding locally, and seeing the changes happen in a
VM.

Also, I set the environment in vagrant so that the output path inside
the VM for webpack is `/usr/local/share/cockpit/subscription-manager`.

Basically, by providing both, a developer can choose to either set up
tooling on the host and use `npm run vagrant-watch` or they can do all
coding in the VM (where `sudo npm run watch` is more appropriate).