Improve the cockpit plugin dev experience
Changes:
 - `/vagrant/src` and `/vagrant/python-rhsm/src` are now in the
   python path for all processes, by using `sitecustomize.py`
 - `/usr/libexec/rhsm-service` is a now symlink to
   `/vagrant/bin/rhsm-service`
 - webpack is now configured the same in the VM
 - `/usr/local/share/cockpit/subscription-manager` is now a symlink to
   `/vagrant/cockpit/dist`
 - Excludes are now adjusted to avoid problems when rsyncing
 - The dbus config file for RHSM1 is replaced with a shim that includes
   `/vagrant/etc-conf/dbus/system.d/com.redhat.RHSM1.conf`
 - selinux is set to permissive to allow the aforementioned shim to work.
 - `nvm` is used to install npm, instead of relying on the version in
   EPEL

With these changes, a developer should be able to simply checkout the
code, `vagrant up`, and see the results in the cockpit interface at ex.
https://centos7.subman.example.com:9090/ . Note that you may need to use
`vagrant destroy` and `vagrant up` to recreate the VM in order to
realize all the effects of the changes in this commit.

Also, the symlinks/shim allow a more interactive D-Bus development, where
only rsync and then stopping the stale process is necessary.