Don't run gui tests by default
Since we're planning on not supporting the GUI w/ Python 3, and they
were causing issues on F27 (segfaults!) that were non-trivial to
determine the root cause of, we'll just stop running them by default.

Ran into these issues while developing #1761, so it may be necessary to
pull these changes in to test that PR against Python 3...

C-I should still be running them by using a different config file for
nose.