1528625: Prevent dmidecode failure from returning None
Calls to `dmidecode.connector()` can fail with a "list assignment index
out of range" error (for reasons that are currently unknown).  In this
case, our code should return an empty dictionary instead of None since
the return value is dereferenced later where it is assumed to be a
dictionary.

Additionally, if there are failures during fact gathering, we should
display the stacktrace in the logs for debugging purposes.