Adds a new cli utility 'intentctl'
This utility maintains the contents of the
/etc/rhsm/intent/intent.json file per user input.
This effectively allows for a user to set the intention for
use (and other aspects such as SLA, and offering) for a given
system.

The source for the separate package is in ./src/intentctl
The tests are in ./test/intentctl
The setup.py file and other packaging files are in ./packages/intentctl

In addition, this adds a new tito VersionTagger to allow us to
maintain the same version across multiple setup.py files.

Additional changes below:

* Add new tagger to tag sub-python packages
* Adds support for unittesting intentctl individually
* Adds tests for the utility methods
* Updates top-level readme to refer to sub-packages