Adds DirectoryWatch class
This class represents a watch on a set of paths on the file system.
It is meant to be a wrapper useful to have a common means of
specifing which files need to be watched.

The flags used for a mask in DirectoryWatch are presently shared
with the pyinotify library (for compatibility).