Make build_ext a proper dependency.
The modules in build_ext are used both by subscription-manager and
syspurpose.  Move it to its own directory and have both
subscription-manager and syspurpose install it from that directory as an
editable dependency.  Also have subscription-manager install syspurpose
as an editable dependency since it has tests that rely on it.

An editable dependency is an equivalent to setuptools' "development
mode".  Instead of the py files being copied to site-packages, a
.egg-link file is laid down instead.  That egg-link file points to the
actual location of the py files so that any changes made to the files in
the dependency are immediately reflected in anything relying on that
dependency.  See
https://setuptools.readthedocs.io/en/latest/setuptools.html#development-mode
for more information.

Also load build_ext modules by munging sys.path.  Treating build_ext as
a dependency to install requires that pip install run first.  Some of
our Jenkins jobs do not run pip install.  Likewise, I don't know that
running pip install would be a viable option in an rpmbuild although we
could just set PYTHONPATH.