1614943: Fix bytes/unicode handling of dmi data
The message "SMBIOS implementations newer than version 2.7 are not fully
supported by this version of dmidecode." is benign: it only means that
information that was added in later revisions of SMBIOS 2.x won't be
accessible via the python dmidecode library.

The real issue was our handling of bytes/strings not working properly in
Python 3.

Easiest was to see the difference is `sudo subscription-manager facts`
on Fedora (can test this change via `sudo PYTHONPATH=./src
subscription-manager facts`).