1633264: Sync changes with Entitlement Server from both subman and syspurpose
This patch helps to unify the code paths used to perform the syncing
of local syspurpose values with those from the connected server.
With this, both the syspurpose cli tool and the subscription-manager
cli tool, will - if the system is registered - sync the local changes
with those made remotely.

This also fixes the following bugs:
  * https://bugzilla.redhat.com/show_bug.cgi?id=1641577
  * https://bugzilla.redhat.com/show_bug.cgi?id=1633264
  * https://bugzilla.redhat.com/show_bug.cgi?id=1625293