Add an environment variable to disable package profile reporting
Adds the SUBMAN_DISABLE_PROFILE_REPORTING environment variable that
supersedes the report_package_profile setting in rhsm.conf in some
cases.  If the environment variable is set to a truth-y value, the value
in rhsm.conf will be ignored.  In all other cases, the value in
rhsm.conf will be used.