1641577: Use cached values if server is unavailable
Using cached values as the values from the server (if all we've got
is an empty dictionary) allows us to make stable modifications to
the local file (in other words, to not remove all other attributes
of syspurpose in the event that the server has nothing for us).

This commit also suppresses the syspurpose tool warning that it is
unable to send syspurpose to subscription management server when
the system appears not to be registered as it makes no sense to
display such a message in that case.