1633264: Ensure we sync syspurpose on register
Also ensure that we properly handle the case where
local syspurpose exists and the server does not support syspurpose.
In this case we should only update the values locally.

In the case that the server spontaneously does support it,
we should not be clearing the values already set locally.
This should also fix: RHBZ 1641577
* https://bugzilla.redhat.com/show_bug.cgi?id=1658389