1658409: Stop redhat.repo from growing exponentially
Thanks long-lived instances of the RepoFile class, subscription-manager
was reading the underlying file twice.  The duplicate reads resulted in
the number of comment nodes doubling in the instance and then getting
written out when it was time to save the file.  Every write of the
redhat.repo file culminated in a doubling of the comment nodes leading
to an exponential growth in file size.

This patch fixes the issue but instantiating the RepoFile class late
into the process and then letting the object get GCed at the end of the
perform method in RepoUpdateActionCommand.