1673973: Do not override sla on auto-attach
Also: Do not hold onto old consumer uuid in dbus
We are not passing anything useful in terms of service level
in to the AutoAttach function. Passing an empty string as we
had been prior to this commit was causing us to override whatever
was set by the user for syspurpose.