Anaconda addon: setup() and execute() no longer get instclass
Since anaconda commit
https://github.com/rhinstaller/anaconda/commit/78fd1e82 ,
anaconda addons should no longer expect an install class to be
passed to their setup() and execute() methods. Without this fix,
the addon breaks recent versions of anaconda. As the policy
for this tool seems to be that the most recent version is sent
to all Fedora releases (and RHEL releases?), I implemented this
in such a way that it will work with both old and new anaconda.

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=1713109

Signed-off-by: Adam Williamson <awilliam@redhat.com>