Send package profiles after updating repositories
When registering a client to a satellite with an activation key,
package lists for a client on the satellite was updated as expected.
However, "Installable Updates" on the satellite page was not updated.
It seems a bug.

This happens because subscription-manager reported the repository
informations of the client before updating repositories as specified in
the activation key. In this case, subscription-manager reported an empty
repository informations. The empty repository informations didn't
contribute to the "Installable Updates".

The reporting should be done after updating repositories.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>