Ensure serial existence before comparison
When comparing for equality (or not equality) with certificates
try to use the 'serial' attribute of the other object.
Check for the 'serial' attribute first and require their values
to be equal for the certificates to be equal (likewise for ne).
This allows us to ensure we are comparing something equitable.