Add --no-insights option; ENT-2471
Insights client plans to implement a process to watch the consumer cert
directory and perform registration automatically when we land a consumer
cert in this directory.

In order to disable this functionality during registration, this change
adds the `--no-insights` option, which disables the planned systemd
unit.

If the systemctl command fails to disable the unit, we proceed with
registration. There may be some use cases (e.g. container where systemd
does not run) where failure to mask may be a non-issue.

This change also looks for a disabled unit and removes it on
unregistration if present.

To test
-------
Try any given register command, appending `--no-insights` option.
Observe that after registration
`systemctl status insights-register.path` explains that the unit is
masked. Note you don't actually have to have a pre-existing
`insights-register.path`, so no need to install insights-client or mock
up the unit.

Afterwards, try `subscription-manager unregister` and observe that the
unit is no longer masked via `systemctl status insights-register.path`.