1848636, 1849074: Update insights machine-id path
Prior the insights-client machine-id (the id used to uniquely
identify a system for Red Hat Insights) was located in the filesystem
at '/etc/redhat-access-insights/machine-id'.
It is (as of 3.0.13+ of insights-client) moved to
'/etc/insights-client/machine-id'.

Subscription-manager needs to report this in order to ensure we
do not end up with two records for the same system (one created
during registration using insights-client and another created
in hosted based on the new consumer registration to RHSM)

Now subscription-manager will first try to read the machine id:
1) from the file specified in insights_client.constants.machine_id_file
2) from "/etc/insights-client/machine-id"
3) from "/etc/redhat-access-insights/machine-id"

Also added a new test function to allow mocking of an arbitrary number
of files on the file system called open_mock_many.