Support Insights registration by default in cockpit
This makes the following changes:
- checks the "Connect this system to Red Hat Insights" option by default
- passes the no_insights option through when the option is unchecked
- reduces any errors encountered during Insights registration to
  warnings

Note I wrapped the use of `cockpit.spawn` in one place to adapt
cockpit promises to standard Promises, this was necessary to get error
logic working properly.

To test
-------
Try a centos 7 machine without insights-client packages available. Not
that by default, you'll see a warning about not being able to install
the insights-client package.

Try unchecking the box and observe the masking introduced in ENT-2471
happens.

Try unregistering via cockpit and observe the registration is unmasked.

Note that there may be some additional cleanup to do here (e.g. may not
be necessary to have cockpit code call `insights --register` in the
future.