improve the help message for attach --auto
There's a discrepancy between the man page and attach --help. The man page
for attach --auto says that it:

"Automatically attaches the *best-matched* compatible subscription or
subscriptions to the system."

However, attach --help omits the word "best-matched". Today, my colleague
wrongly assumed from --help that subscription-manager attaches all
compatible subscriptions with the --auto option. This is not true though.
It indeed attaches only the best-matched ones. This commit fixes the confused
--help text by adding the "best-matched" word to it.

Thanks Lars Karlitski for pointing this out!