1886772: Add in memory read through cache, delete SCA cert when not needed
This stops bad cache values from incorrectly displaying that
the system is registered to an organization which is running
in Simple Content Access mode.

Now the organization's content access mode is cached only in memory
so as to limit the duration of it's validity to the lifetime of
any of our applications runtime.