cockpit: Lower cockpit-ws dependency to Recommends
Cockpit web UI pages should not depend on the cockpit web server. It is
perfectly reasonable and often desirable to not run the web server on a
managed machine, and log in through the host switcher and SSH instead.

However, in this case the dependency is there because of the cockpit
desktop integration. We mostly care about the default OS installation
there, so lower it to Recommends (which are installed by default) and
guard the .desktop file with a `TryExec=`.

(cherry picked from commit 65c75c6d8a958a20529020bef03a9ed612d7ddd9)