Rename internal variable for syspurpose --show
By default, a '--show' option in ArgumentParser is saved as a namespace
variable 'show'. While it is currently not a problem for the '--show'
option of the 'syspurpose' command, it will be a problem when adding the
options of the other syspurpose subcommands as subparsers of the
'syspurpose' command.

Hence, bind the ArgumentParser option result to a less generic variable
name to avoid clashes.