1976225: read lscpu from its JSON output if available
lscpu has a -J/--json option to output the data as JSON since
util-linux 2.30. Since it is obviously meant for machine parsing, and
the "human output" recently changed in ways that broke our simple
parser, then switch to the JSON output.

Query lscpu to check whether it has a --json parameter; this is done
scraping the --help text, which is not nice, however should do it;
if --json is available, invoke lscpu that way, and parse the JSON output
to populate the facts from it.

Card ID: ENT-4040