1876828: Try to suppress errors in stderr when not run as root
* BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1876828
* BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1890314
* BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1876841
* BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1947844
* Card ID: ENT-3051
* Card ID: ENT-3852

Before this change the only log location was /var/log/rhsm/rhsm.log. If
the subscription-manager was run as non-root user (as dnf plugin, e.g.
'dnf list'), it sent several error messages to stderr, polluting the
output.

Now the UID of the subman process is checked:
- If the program is run as root (UID 0), the log location is still
  /var/log/rhsm/rhsm.log.
- If the program is run as non-root (other UIDs), the log location is
  $XDG_CACHE_HOME/rhsm/rhsm.log (~/.cache/rhsm/rhsm.log in most cases).
- If the location is not writable, the errors will be sent to stderr
  as they were before this change.

This behavior is now mentioned in the man page.
/var/log/rhsm/rhsm.log has been added to the man page's FILES section.