Drop YUM plugins
YUM has not been used in Fedora since 22, and in RHEL since 8; because
of that, the plugins for YUM have not been worked on and tested for a
long time.

Hence, drop the YUM plugins, as they are essentially dead code at this
point. Remove also the only configuration file that is not used by the
dnf plugins, and the associated tests.

Card-ID: ENT-4160