Drop RHN migration
It was used to migrate from the old Red Hat Network to Red Hat
Subscription Management. This migration was completed few years ago, and
it has not been supported anymore for a couple of years.

As it was specific to RHEL, and RHEL 9 does not even ship these
bits (deprecated in RHEL 8 already) anymore, then simply drop the
migration code.

Card-ID: ENT-4155