Adding toolbox containers for development and test
This adds fedora based toolbox containers and
updates the jenkinsfile of the project so that
tests can be run in a similar environment both
on jenkins nodes and locally.

This allows test failures to be reproduced more
easily. See the included readme updates for
details on the operation of the containers.

The containers can also be used for development
as well.

Allow toolbox create to pull images automatically