Update exception handling
* Card ID: ENT-4228

ExceptionMapper's "get_message" returned "None" if supplied exception
was not known to the class. Now, a string representation is returned
instead.

This change will allow us to use the mapper for all the exceptions, not
just the ones defined in there.

Because of this change some method and variable names were updated to
closely describe their meaning and usage.