1983144: More useful feedback on unknown argument
* BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1983144
* Card ID: ENT-4233

When unknown subcommand or argument is supplied, an error message is
printed and sub-man exits. This commit changes the error message to be
the same as it was before we switched from optparse to argparse.

(cherry picked from commit 8075c66adb8bb4fbbcd2c173b8f1c18f71a6e4df)