1859157: Display better error message on incorrect --org
* BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1859157
* Card ID: ENT-2701

If the user has more than one organization the --org argument can be
used to specify one of them.

If the --org value was set to some invalid string, error message was
just a generic text that the list of valid values could not be obtained.

Now the proper message, returned by the server, is shown and EX_SOFTWARE
exit code is used.