2003777: Show available organizations before asking for input
* BZ: https://bugzilla.redhat.com/show_bug.cgi?id=2003777
* Card ID: ENT-4316

If the user has more than one organization available and uses --list
to show one of system purposes they are asked to enter key of one of
their organizations.

Now, before they are prompted to enter the organization, a list of all
available organizations will be shown.

This functionality is already available in the 'register' command since
1ad8b33 (2020-02).