ENT-4340: Resolve deprecation warnings
* Card ID: ENT-4340

This commit resolves some of the deprecation warnings that are being
emitted by parts of subscription-manager code. They can be seen, for
example, when 'pytest' is run.

Following changes have been made:

- 'assertRaisesRegex()' instead of 'assertRaisesRegexp()',
- 'assertEqual()' instead of 'assertEquals()',
- 'warning()' instead of 'warn()',
- 'bind_textdomain_codeset() is being removed.

The 'bind_textdomain_codeset' has been deprecated since Python 3.8 and
is being removed in Python 3.10. It only affects calls to 'l'-prefixed
functions of gettext ('lgettext()') which are not used in the code base.