2003777: Only hint organizations if it's needed
* BZ: https://bugzilla.redhat.com/show_bug.cgi?id=2003777
* Card ID: ENT-4411

Recent patch added a hint message to the --org argument. This update
slightly changes the behavior, so the hint is only displayed when the
user did not use the --org at all.

When they used it and it's valid, the hint is skipped and the syspurpose
subcommand is run as usual.

When they used it and it's invalid, the hint is skipped and formatted
RhsmlibException is printed saying that the organization could not be
found.

Now the behavior will be the same as it is in the register subcommand.

(cherry picked from commit ab76b029d9cad5fdff8c36f68b9e6517b9033e19)