cockpit: use the self-signed key in mock-insights
Make sure that the mock-insights tool properly uses also the self-signed
key created by cockpit, not only the certificate. The only gotcha is due
to a behaviour change in cockpit: v242 switched from a single file
including both key and certificate to separate files for them.

Credits to both Matej Marusak and Martin Pitt (of the cockpit team) for
the hints on this.