cockpit: fix system installation of subscription-manager
It seems that newer versions of Python or setuptools create a different
layout in the Python site-package depending whether the installation is
direct to the specified prefix, or there is temporary location (a local
destdir). Because of this, force "--root /" when installing directly to
/usr for testing: this seems to use the same layout of the system
installation (via distribution packages), and thus the existing files
will be properly overwritten.

Another change is the location of the "subscription-manager" executable:
setuptools installs it in $prefix/bin, while our Makefile moves it to
$prefix/sbin; hence, move it to the desired place, so it properly
overwrites the package version.

(cherry picked from commit c7d74b75bdbd55cd1d1f760a894f154fbf293c1c)