cockpit: skip RHEL 9 tests using insights-client
It seems that currently insights-client on RHEL 9 is broken:

  Traceback (most recent call last):
    File "/usr/lib/python3.9/site-packages/insights_client/run.py", line 4, in <module>
      from insights import package_info
  ModuleNotFoundError: No module named 'insights'

Hence, disable the two tests using it when running the test suite on a
rhel-9-X image.

This is mostly a temporary solution to get the cockpit test suite
running on RHEL 9; this commit can be reverted once insights-client in
RHEL 9 is functional.