cockpit: validate the SSL connection to candlepin
As the services image of the cockpit CI now exports the Candlepin CA
certificates, import them in the cockpit machine, and make them
available as system CAs. This allows the connections done with e.g.
python-requests to be validated.

In addition, switch the Candlepin address to the service machine
hostname: this is because the certificate has a CN with the hostname,
so the destination address of the connection must match that.

Also, import the Candlepin CA certificates also in the own store of
subscription-manager: this allows subscription-manager to properly
validate the connection, thus removing the need to set insecure=1 in
rhsm.conf.