2015173: chmod /etc/pki/entitlement/*.pem only when existing
There are no .pem files in /etc/pki/entitlement in case a system is not
registered; in that case, the shell expansion will error out, producing
an error message (ignored) during upgrades.

Instead, find the files manually and run chmod on whatever was found;
this way there is no behaviour change (the permissions of .pem files
are properly set), there is no message in case of no certificates, and
(bonus) the errors from chmod are no more ignored.

Card ID: ENT-4473