tests: drop test_po_files.py
This test used to do some comprehensive testing of the translations,
long time ago. Nowadays, it does few things:
1) tries to translate a certain message in some languages
2) checks that the translation of a different message is unicode

First of all: these tests load translations from the system, so they
should be marked as functional.

Second: gettext() returns the same string in case the string is not
translated in the current locale, so in some cases the original strings
are checked to be unicode, which they obviously are.

Third: gettext (and Weblate, which is currently used to manage
translations) already do a good job in checking for wrongly encoded
messages, so there is no need to do the same check locally.

Hence, simply drop test_po_files.py, as there is nothing useful in it.

Card ID: ENT-4484

(Cherry-picked from e9cb4bc)