Ensure datetime.timezone.utc objects for parsed UTC dates
The UTC tzinfo objects in datetime.datetime objects returned by
dateutil.parser.parse() are dateutil's own: because of this, checks for
datetime.timezone.utc as tzinfo in Python standard library modules fail.
Hence, switch the tzinfo object for UTC datetime's to
datetime.timezone.utc.

Adapt the isodate test suite for that:
- check that the tzinfo is datetime.timezone.utc
- datetime.timezone.dst() always return None (documented as such), so
  drop that check

Adapt also the test_connection.py to the new expected timezone for the
datetime passed as 'if_modified_since'.