2026320: fix format of HTTP-date headers
Use email.utils.format_datetime() to format the datetime for HTTP-date
headers: unlike email.utils.formatdate() it works with
datetime objects, so there is no need to do a buggy roundtrip though
time objects.

This fixes the formatting of the 'If-Modified-Since' in particular.

Rework test_date_formatted_properly_with_japanese_locale() with parts
from TestContentAccessCache to check a bit more that 'If-Modified-Since'
is properly formatted when invoked by ContentAccessCache.

Card ID: ENT-4538