2026286: consider user-specified --org in any case
Right now, commands that can take --org ignore its value in case the
user belongs to only a single organization; this means that wrong values
passed by the user are ignored in that case, unlike what happens in case
the user belongs to more than one organization.

As a way to unify the behaviour in all the cases, always accept the
--org specified by the user, trying the automatic setting or the user
query otherwise.

This also saves one call to the server in case --org is specified.

Extend the tests of RoleCommand to verify this behaviour.