connection: improve the internal ProxyException
Make it carry the proxy hostname & port, and optionally also the
exception that caused the issue. This way it will be possible to provide
a better diagnostic for it.

Adapt the various places that raise it to provide all the attributes
they know about.

(cherry picked from commit a4e6aa426fdce5ce229b0109dd18376e72d48fc1)