Improve formatting of ProxyException
In case there is any proxy related error as ProxyException, format it
properly using the details available in the ProxyException itself.
In particular, always print the string representation of the issue, with
also the error code in case of socket issues.

Make use of ExceptionMapper to actually use this formatting of
ProxyException on errors, replacing the former generic message.