Squelch DeprecationWarning's when importing pyinotify
pyinotify uses asyncore, which was deprecated in Python 3.6. Since
pyinotify has been (at the time of this writing) not active for 6 years
and more, we will probably not get a newer version switching away from
asyncore soon.

Hence, apply a localized hammer: temporarily disable the
DeprecationWarning warnings when importing pyinotify.