2023430: cockpit: improve handling of main curtain view
The logic used to determine what to show as curtain view (i.e. before
all the subscription data & status were loaded) was partially flawed:
if a valid subscription status was determined not as very last data
at the very end of the loading phase, then the curtain view briefly
showed the "Unable to connect" error message, just to be quickly
replaced by the actual subscription view. This happended because the
logic used in SubscriptionsView.renderCurtains() did not know the valid
subscription statuses.

Shuffle the logic a bit to more to avoid such mistakes:
- move the check for the "Updating" view as first, using it when either
  it is loaded w/o a valid state, or w/ a valid state and not fully
  loaded; both situation are correct, as they represent that only some
  data was fetched (and the rest is still being loaded)
- explicitly check for all the currently (as of sub-man 1.29.23) valid
  statuses for subscriptions; any new status in the future can simply be
  added to the list of checked states, and its lack will simply show the
  "Unable to connect" view again
- enhance the "Unable to connect" EmptyStatePanel to show also the
  status and its UI string; while we could easily log them to console,
  having the not known status printed to the screen will greatly help
  debugging such issues in the future, especially when reported without
  the open browser console
- also add the "service-unavailable" status among the ones that force
  the curtain view; it is already checked in renderCurtains()