2046302: Apply Conscious language initiative changes
* BZ: https://bugzilla.redhat.com/show_bug.cgi?id=2046302
* Card ID: ENT-4680

This is a partial backport of ENT-3764 (5b6db92), as it only changes the
words that are visible to the users: config files and man pages.

For maintenance and testing reasons no other changes have been made:
variable names and comments were not updated and have been kept the
same.