environments: fix list for account in env-less org
In case an account registers to an organization without environments on
a Candlepin with "environments" capability, the list of environments for
it is null. Hence, fall back to an empty string, so that
"environments --list" properly reports the "no environments to report"
message rather erroring out because "NoneType has no len()".