yum/dnf plugins: properly check for SCA
Properly check the status of SCA to using the helper function for it,
instead of having inline code that will break the flow of the function
on return. This also aligns these changes in the
subscription-manager-1.24.13 branch with the subscription-manager-1.24
branch.

Fixes commit 944756e8a4daf27f7339a1c6e8d182b6112c08a3.