cockpit: set $RPM_BUILD_ROOT w/ installing
Due to a recent change in Fedora 36 [1], the installation prefix is
mangled in certain scenarios [2].

Hence, set a $RPM_BUILD_ROOT environment variable when invoking
'setup.py install', so it works again as it used in older versions of
Fedora. This script is used only for the cockpit CI testing, so there
is no change for usual builds.

[1] https://src.fedoraproject.org/rpms/python3.10/pull-request/63
[2] https://bugzilla.redhat.com/show_bug.cgi?id=2026979