integration-test: Use custom TLS certs for mock-insights
This makes it work again on rhel-9-0.

There is a bug in sscg on rhel-9-0 that causes Cockpit to create its
certificates with openssl.  However, those don't work for
mock-insights, so let's just make a custom set just for it.

Also, use file locations and port numbers that are allowed by SELinux
on rhel-9-0.