Add Black to CI
* Card ID: ENT-4847

The version is pinned for consistency. If you have different version
than the Jenkins does (because one of the systems was not updated while
the other one was), the build will fail even though your local Black
claims everything is in order.

For that reason it would be the best to use the version from PyPI via
pip and not system package version from repository, which may be out of
sync.

If this situation does happen and Black in CI finds some issues, it will
show a diff for the changes, so it can be fixed even manually. But the
best-case scenario is seamless check with the versions synchronized.