2074111: clear the release cache on release change
Commit dd6eba0b1d6722d06967e1a7e8e65ccb9ba00858 fixed the usage of the
status cache, making sure it is properly used in more cases. One side
effect it happened was that the release status cache was used as it was,
and since nothing invalidates it, any release change (done with
`release --set/--unset`) was never reflected.

Hence, whenever the release is successfully changed (via
updateConsumer()), delete the release status cache: the cache will be
refreshed whenever there is the need for it (e.g. when trying to expand
"$releasever" in repository URLs). Remove the cache before
RepoActionInvoker.update() is invoked, as it will update the
repositories, potentially asking for the release version to expand.

(cherry picked from commit 61fed7d4dff6639697db9e1b0dc7ec8ed4c1e864)