cockpit: test with split subscription-manager-cockpit
The cockpit plugin was recently split in its own repository, and will be
soon removed from this repository; nevertheless, we still want to test
subscription-manager from sources with subscription-manager-cockpit,
to make sure changes in sub-man do not break the cockpit plugin.

To do this, reuse the testing infra of subscription-manager-cockpit as
much as possible: let it setup the test image with the cockpit plugin,
and then install sub-man from the sources of this repository. The tests
run are those in subscription-manager-cockpit, so there is only a single
place with them.