Run CI's pytest with verbose flag
When tests produce some unexpected output ("write to closed file") or
print a traceback, it is difficult to pinpoint which of them is the
cause, because we're randomizing their order every time.

By using verbose mode, we'll see exactly which test is misbehaving and
we can skip the part where we have to go looking for it.